backend/pyproject.toml
[tool.poetry]
name = "spiritual-qa-backend"
version = "0.1.0"
description = "Backend for Spiritual Q&A Platform using RAG"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
argon2-cffi = "^23.1.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.28"
qdrant-client = "^1.8.0"
openai = "^1.13.3"
llama-index = "^0.10.18"
python-multipart = "^0.0.9"
alembic = "^1.13.1"
redis = {extras = ["hiredis"], version = "^5.0.0"}
email-validator = "^2.1.0"
python-dotenv = "^1.0.0"
pypdf = "^4.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"
ruff = "^0.3.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=85"
markers = [
    "integration: marks tests requiring Docker services",
]
